working directory works
